Fix CRI image store cache inconsistency
The CRI image store maintains a `refCache` (map[string]string) that
stores image refs -> image digests.

When an image is stored in the CRI image store, if the ref is not
normalized (missing a Docker registry prefix), it will still be stored
as-is in the `refCache`. However, the references set on the image object
being stored *will* be normalized. This cache inconsistency causes
issues when removing images with not normalized names (e.g.
"busybox:latest"), because the cleanup logic removes the refs in
`image.References` (normalized) from `refCache` (not normalized).

To fix, ensure any writes to `refCache` use parsed image references.

Signed-off-by: Chris Henzie <chrishenzie@google.com>