WIP: Replace legacy NRI client in PodSandbox controller
The idea here is to replace the legacy client with the latest NRI client
used elsewhere in the CRI server code.

The part I'm unsure about is how I get the sstore.Sandbox to pass to the
NRI client. I'm building a dummy sstore.Sandbox in the meantime but that
doesn't seem right.

This is likely broken in a variety of ways.